requirements.txt
streamlit
pandas
faiss-cpu
numpy
sentence-transformers
huggingface_hub
PyMuPDF
